Don't assume bgzip is on path when tabix is installed.
I encountered a bug where pybedtools gives an error message
about /tmp/something.gz not existing when trying to tabix
a bed file. This turned out to be because bgzip was missing
from the path (but tabix was present). This checks for both
binaries separately, and results in a more helpful error
message.