cat: handle empty input files
The latest release (0.7.2) has an issue with bedtools cat on empty files
due to fixes for #138 (https://github.com/daler/pybedtools/commit/08a07378200906dbc75c2c87413d56e9bad8ff94).
Passing an empty file to cat will result in field_nums having a 0, and
the new minfields check will eliminate all of the fields in the
non-emtpy files.

This has the small fix for that plus a test case to catch it.