Ignore bedtools warnings that are not capitalized
Bedtools does not have uniform way of logging warnings. This results
that some warning messag look like so:

    WARNING: Foo bar...

and others like so:

    Warning: Foo bar...

Pybedtools is currently ignoring warnings that start with "WARNING"
but not the ones that start with "Warning". This inconsistency causes
that some warnings are converted to errors which is undesired.